pyproject.toml
# Project configuration for Agentic Trek

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-trek"
version = "1.0.0"
description = "A modern recreation of the classic Trek game with intelligent AI agents"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Python Developer", email = "developer@example.com"}
]
keywords = ["game", "ai", "trek", "strategy", "pygame"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment :: Turn Based Strategy",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    "pygame>=2.6.1",
    "numpy>=1.24.3",
    "scipy>=1.10.1",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.3",
    "pyyaml>=6.0.1",
    "colorama>=0.4.6",
    "click>=8.1.6",
    "tqdm>=4.65.0",
    "jsonschema>=4.19.0",
    "psutil>=5.9.5",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-benchmark>=4.0.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.1",
    "isort>=5.12.0",
    "pre-commit>=3.3.0",
    "sphinx>=7.1.2",
    "sphinx-rtd-theme>=1.3.0",
]
docs = [
    "sphinx>=7.1.2",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.24.0",
]
analysis = [
    "matplotlib>=3.7.2",
    "seaborn>=0.12.2",
    "jupyter>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/agentic-trek"
Repository = "https://github.com/yourusername/agentic-trek"
Documentation = "https://agentic-trek.readthedocs.io"
"Bug Tracker" = "https://github.com/yourusername/agentic-trek/issues"

[project.scripts]
trek = "src.main:main"

# Tool configurations

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pygame.*",
    "numpy.*",
    "scipy.*",
    "sklearn.*",
    "pandas.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "ai: marks tests as AI-specific tests",
    "performance: marks tests as performance tests",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg-info",
    ".venv",
    ".tox",
]
per-file-ignores = [
    "__init__.py:F401",
    "tests/*:D",
]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101", "B601"]
